requirements.txt
Flask
flask-restx
marshmallow
Flask-SQLAlchemy
Flask-Admin
pytest
